Renamed triggerv2 to trigger.
