+++ internal/trigger/.gitignore
