+++ internal/trigger/postgres_query_trigger_test.go
@@ -1,4 +1,4 @@
-package triggerv2
+package trigger
 
 import (
 	"context"
@@ -22,24 +22,16 @@ import (
 	"github.com/zclconf/go-cty/cty"
 )
 
-func TestTriggerQuery(t *testing.T) {
+func TestPostgresSqlTriggerQuery(t *testing.T) {
 	ctx := context.Background()
 
 	cache.InMemoryInitialize(nil)
 
 	assert := assert.New(t)
 
-	sourceDbFilename := "./test_trigger_query.db"
-	_, err := os.Stat(sourceDbFilename)
-	if !os.IsNotExist(err) {
-		err = os.Remove(sourceDbFilename)
-		if err != nil {
-			assert.Fail("Error removing test db", err)
-			return
-		}
-	}
+	connectionString := "postgres://flowpipe:password@localhost:5432/flowpipe-test?sslmode=disable"
 
-	db, err := sql.Open("sqlite3", sourceDbFilename)
+	db, err := sql.Open("postgres", connectionString)
 	if err != nil {
 		assert.Fail("Error initializing db", err)
 		return
@@ -64,7 +56,13 @@ func TestTriggerQuery(t *testing.T) {
 		return
 	}
 
-	err = createTestTableA(db, "test_one")
+	err = dropPostgresSqlTestTable(db, "test_one")
+	if err != nil {
+		assert.Fail("Error deleting test table", err)
+		return
+	}
+
+	err = createPostgresSqlTestTableA(db, "test_one")
 	if err != nil {
 		assert.Fail("Error creating test table", err)
 		return
@@ -94,7 +92,7 @@ func TestTriggerQuery(t *testing.T) {
 		},
 	}
 
-	err = populateTestTableA(db, "test_one", data)
+	err = populatePostgresSqlTestTableA(db, "test_one", data)
 	if err != nil {
 		assert.Fail("Error populating test table", err)
 		return
@@ -152,7 +150,7 @@ func TestTriggerQuery(t *testing.T) {
 	}
 
 	trigger.Config = &modconfig.TriggerQuery{
-		Database:   "sqlite:./test_trigger_query.db",
+		Database:   connectionString,
 		Sql:        "select * from test_one",
 		PrimaryKey: "id",
 		Captures: map[string]*modconfig.TriggerQueryCapture{
@@ -176,8 +174,11 @@ func TestTriggerQuery(t *testing.T) {
 		return
 	}
 
-	assert.Equal(1, len(pipelineQueues), "wrong number of pipeline queues there should be only 1 insert")
-	assert.Equal("insert_pipe", pipelineQueues[0].Name, "wrong pipeline name")
+	// The callback to the mocks should have been called by now
+	if generatedEvalContext == nil {
+		assert.Fail("generated eval context should not be nil")
+		return
+	}
 
 	selfVar := generatedEvalContext.Variables["self"]
 	if selfVar == cty.NilVal {
@@ -214,6 +215,10 @@ func TestTriggerQuery(t *testing.T) {
 		}
 	}
 
+	// check the triggerCommands .. we check how many pipeline is executed and which pipeline is executed
+	assert.Equal(1, len(pipelineQueues), "wrong number of trigger commands only the insert pipeline should be executed")
+	assert.Equal("insert_pipe", pipelineQueues[0].Name, "wrong pipeline name")
+
 	//
 	// SECOND RUN
 	//
@@ -228,7 +233,7 @@ func TestTriggerQuery(t *testing.T) {
 		return
 	}
 
-	assert.Equal(0, len(pipelineQueues), "pipeline queues should be nil, since there's no change the pipeline should NOT be called")
+	assert.Equal(0, len(pipelineQueues), "trigger command should be nil, since there's no change the pipeline should NOT be called")
 
 	//
 	// THIRD RUN
@@ -252,7 +257,7 @@ func TestTriggerQuery(t *testing.T) {
 		},
 	}
 
-	err = populateTestTableA(db, "test_one", data)
+	err = populatePostgresSqlTestTableA(db, "test_one", data)
 	if err != nil {
 		assert.Fail("Error populating test table", err)
 		return
@@ -267,7 +272,6 @@ func TestTriggerQuery(t *testing.T) {
 		return
 	}
 
-	assert.NotNil(pipelineQueues, "trigger command should not be nil")
 	// The callback to the mocks should have been called by now
 	if generatedEvalContext == nil {
 		assert.Fail("generated eval context should not be nil")
@@ -302,7 +306,7 @@ func TestTriggerQuery(t *testing.T) {
 		}
 	}
 
-	assert.Equal(1, len(pipelineQueues), "wrong number of pipeline queue, only the insert pipeline should be executed")
+	assert.Equal(1, len(pipelineQueues), "wrong number of trigger commands only the insert pipeline should be executed")
 	assert.Equal("insert_pipe", pipelineQueues[0].Name, "wrong pipeline name")
 
 	//
@@ -310,7 +314,7 @@ func TestTriggerQuery(t *testing.T) {
 	//
 	// Test for update
 
-	err = updateTestTableA(db, "test_one", map[string]interface{}{
+	err = updatePostgresSqlTestTableA(db, "test_one", map[string]interface{}{
 		"id":                "1",
 		"name":              "John",
 		"age":               35,
@@ -366,7 +370,7 @@ func TestTriggerQuery(t *testing.T) {
 		}
 	}
 
-	assert.Equal(1, len(pipelineQueues), "wrong number of pipeline queue commands only the update pipeline should be executed")
+	assert.Equal(1, len(pipelineQueues), "wrong number of trigger commands only the update pipeline should be executed")
 	assert.Equal("update_pipe", pipelineQueues[0].Name, "wrong pipeline name")
 
 	//
@@ -390,7 +394,7 @@ func TestTriggerQuery(t *testing.T) {
 	//
 	// Delete some rows
 	idsToDelete := []any{"1", "4"}
-	err = deleteFromTestTable(db, "test_one", idsToDelete)
+	err = deleteFromPostgresSqlTestTable(db, "test_one", idsToDelete)
 	if err != nil {
 		assert.Fail("Error deleting from test table", err)
 		return
@@ -461,14 +465,14 @@ func TestTriggerQuery(t *testing.T) {
 		},
 	}
 
-	err = populateTestTableA(db, "test_one", data)
+	err = populatePostgresSqlTestTableA(db, "test_one", data)
 	if err != nil {
 		assert.Fail("Error populating test table", err)
 		return
 	}
 
 	idsToDelete = []any{"2"}
-	err = deleteFromTestTable(db, "test_one", idsToDelete)
+	err = deleteFromPostgresSqlTestTable(db, "test_one", idsToDelete)
 	if err != nil {
 		assert.Fail("Error deleting from test table", err)
 		return
@@ -483,30 +487,22 @@ func TestTriggerQuery(t *testing.T) {
 		return
 	}
 
-	assert.Equal(2, len(pipelineQueues), "wrong number of commands")
+	assert.Equal(2, len(pipelineQueues), "wrong number of trigger commands only the delete pipeline should be executed")
 	assert.Contains([]string{"delete_pipe", "insert_pipe"}, pipelineQueues[0].Name, "wrong pipeline name")
 	assert.Contains([]string{"delete_pipe", "insert_pipe"}, pipelineQueues[1].Name, "wrong pipeline name")
 	assert.False(pipelineQueues[0].Name == pipelineQueues[1].Name, "ensure that we don't call insert_pipe twice or delete_pipe twice")
 }
 
-func TestTriggerQueryNoPrimaryKey(t *testing.T) {
+func TestPostgresSqlTriggerQueryNoPrimaryKey(t *testing.T) {
 	ctx := context.Background()
 
 	cache.InMemoryInitialize(nil)
 
 	assert := assert.New(t)
 
-	sourceDbFilename := "./test_trigger_query.db"
-	_, err := os.Stat(sourceDbFilename)
-	if !os.IsNotExist(err) {
-		err = os.Remove(sourceDbFilename)
-		if err != nil {
-			assert.Fail("Error removing test db", err)
-			return
-		}
-	}
+	connectionString := "postgres://flowpipe:password@localhost:5432/flowpipe-test?sslmode=disable"
 
-	db, err := sql.Open("sqlite3", sourceDbFilename)
+	db, err := sql.Open("postgres", connectionString)
 	if err != nil {
 		assert.Fail("Error initializing db", err)
 		return
@@ -531,7 +527,13 @@ func TestTriggerQueryNoPrimaryKey(t *testing.T) {
 		return
 	}
 
-	err = createTestTableA(db, "test_one")
+	err = dropPostgresSqlTestTable(db, "test_one")
+	if err != nil {
+		assert.Fail("Error deleting test table", err)
+		return
+	}
+
+	err = createPostgresSqlTestTableA(db, "test_one")
 	if err != nil {
 		assert.Fail("Error creating test table", err)
 		return
@@ -561,7 +563,7 @@ func TestTriggerQueryNoPrimaryKey(t *testing.T) {
 		},
 	}
 
-	err = populateTestTableA(db, "test_one", data)
+	err = populatePostgresSqlTestTableA(db, "test_one", data)
 	if err != nil {
 		assert.Fail("Error populating test table", err)
 		return
@@ -619,7 +621,7 @@ func TestTriggerQueryNoPrimaryKey(t *testing.T) {
 	}
 
 	trigger.Config = &modconfig.TriggerQuery{
-		Database: "sqlite:./test_trigger_query.db",
+		Database: connectionString,
 		Sql:      "select * from test_one",
 		Captures: map[string]*modconfig.TriggerQueryCapture{
 			"insert": insertCapture,
@@ -635,8 +637,6 @@ func TestTriggerQueryNoPrimaryKey(t *testing.T) {
 	triggerExecutionId := util.NewTriggerExecutionId()
 
 	triggerRunner := NewTriggerRunner(trigger, executionId, triggerExecutionId)
-
-	assert.NotNil(triggerRunner, "trigger runner should not be nil")
 
 	pipelineQueues, err := triggerRunner.ExecuteTriggerWithArgs(ctx, nil, nil)
 	if err != nil {
@@ -726,11 +726,14 @@ func TestTriggerQueryNoPrimaryKey(t *testing.T) {
 		},
 	}
 
-	err = populateTestTableA(db, "test_one", data)
+	err = populatePostgresSqlTestTableA(db, "test_one", data)
 	if err != nil {
 		assert.Fail("Error populating test table", err)
 		return
 	}
+
+	// Reset
+	generatedEvalContext = nil
 
 	pipelineQueues, err = triggerRunner.ExecuteTriggerWithArgs(ctx, nil, nil)
 	if err != nil {
@@ -787,7 +790,7 @@ func TestTriggerQueryNoPrimaryKey(t *testing.T) {
 	//
 	// Update doesn't work with no primary key
 
-	err = updateTestTableA(db, "test_one", map[string]interface{}{
+	err = updatePostgresSqlTestTableA(db, "test_one", map[string]interface{}{
 		"id":                "1",
 		"name":              "John",
 		"age":               35,
@@ -840,7 +843,7 @@ func TestTriggerQueryNoPrimaryKey(t *testing.T) {
 	deletedKeysSlice := deletedKeys.AsValueSlice()
 	assert.Equal(1, len(deletedKeysSlice), "wrong number of deleted keys")
 	// This Hash is generated in the function RunOne
-	assert.Equal("a7f390faa9ddca021f647b042c2c127f70e49ed3bdec2194df4e784367b5416a", deletedKeysSlice[0].AsString(), "wrong deleted key")
+	assert.Equal("0432ffd1257d9c9e9903528dc8367bed39b5aff0afe68662fd0afa4e85b20b56", deletedKeysSlice[0].AsString(), "wrong deleted key")
 
 	// because update doesn't work without primary key, we have insert & delete instead
 	assert.Equal(2, len(pipelineQueues), "wrong number of trigger commands only the delete pipeline should be executed")
@@ -849,24 +852,14 @@ func TestTriggerQueryNoPrimaryKey(t *testing.T) {
 	assert.False(pipelineQueues[0].Name == pipelineQueues[1].Name, "ensure that we don't call insert_pipe twice or delete_pipe twice")
 }
 
-func TestTriggerQueryB(t *testing.T) {
+func TestPostgresSqlTriggerQueryB(t *testing.T) {
 	ctx := context.Background()
-
-	cache.InMemoryInitialize(nil)
 
 	assert := assert.New(t)
 
-	sourceDbFilename := "./test_trigger_query_b.db"
-	_, err := os.Stat(sourceDbFilename)
-	if !os.IsNotExist(err) {
-		err = os.Remove(sourceDbFilename)
-		if err != nil {
-			assert.Fail("Error removing test db", err)
-			return
-		}
-	}
+	connectionString := "postgres://flowpipe:password@localhost:5432/flowpipe-test?sslmode=disable"
 
-	db, err := sql.Open("sqlite3", sourceDbFilename)
+	db, err := sql.Open("postgres", connectionString)
 	if err != nil {
 		assert.Fail("Error initializing db", err)
 		return
@@ -891,7 +884,13 @@ func TestTriggerQueryB(t *testing.T) {
 		return
 	}
 
-	err = createTestTableB(db, "test_one")
+	err = dropPostgresSqlTestTable(db, "test_one")
+	if err != nil {
+		assert.Fail("Error deleting test table", err)
+		return
+	}
+
+	err = createPostgresSqlTestTableB(db, "test_one")
 	if err != nil {
 		assert.Fail("Error creating test table", err)
 		return
@@ -931,7 +930,7 @@ func TestTriggerQueryB(t *testing.T) {
 		item["blob_data"] = blobData
 	}
 
-	err = populateTestTableB(db, "test_one", data)
+	err = populatePostgresSqlTestTableB(db, "test_one", data)
 	if err != nil {
 		assert.Fail("Error populating test table", err)
 		return
@@ -989,7 +988,7 @@ func TestTriggerQueryB(t *testing.T) {
 	}
 
 	trigger.Config = &modconfig.TriggerQuery{
-		Database:   "sqlite:./test_trigger_query_b.db",
+		Database:   connectionString,
 		Sql:        "select * from test_one",
 		PrimaryKey: "id",
 		Captures: map[string]*modconfig.TriggerQueryCapture{
@@ -1012,7 +1011,6 @@ func TestTriggerQueryB(t *testing.T) {
 		assert.Fail("Error executing trigger", err)
 		return
 	}
-
 	// The callback to the mocks should have been called by now
 	if generatedEvalContext == nil {
 		assert.Fail("generated eval context should not be nil")
@@ -1095,7 +1093,7 @@ func TestTriggerQueryB(t *testing.T) {
 		data[1]["blob_data"].([]byte)[i] = byte(rand.Intn(256)) //nolint:gosec // just a test case
 	}
 
-	err = updateTestTableB(db, "test_one", data[1])
+	err = updatePostgresSqlTestTableB(db, "test_one", data[1])
 	if err != nil {
 		assert.Fail("Error updating test table", err)
 		return
@@ -1161,21 +1159,16 @@ func TestTriggerQueryB(t *testing.T) {
 	}
 }
 
-func TestTriggerQueryBCustomCapture(t *testing.T) {
+func TestPostgresSqlTriggerQueryBCustomCapture(t *testing.T) {
 	ctx := context.Background()
+
+	cache.InMemoryInitialize(nil)
 
 	assert := assert.New(t)
 
-	sourceDbFilename := "./test_trigger_query.db"
-	_, err := os.Stat(sourceDbFilename)
-	if !os.IsNotExist(err) {
-		err = os.Remove(sourceDbFilename)
-		if err != nil {
-			assert.Fail("Error removing test db", err)
-			return
-		}
-	}
-	db, err := sql.Open("sqlite3", sourceDbFilename)
+	connectionString := "postgres://flowpipe:password@localhost:5432/flowpipe-test?sslmode=disable"
+
+	db, err := sql.Open("postgres", connectionString)
 	if err != nil {
 		assert.Fail("Error initializing db", err)
 		return
@@ -1200,7 +1193,13 @@ func TestTriggerQueryBCustomCapture(t *testing.T) {
 		return
 	}
 
-	err = createTestTableB(db, "test_one")
+	err = dropPostgresSqlTestTable(db, "test_one")
+	if err != nil {
+		assert.Fail("Error deleting test table", err)
+		return
+	}
+
+	err = createPostgresSqlTestTableB(db, "test_one")
 	if err != nil {
 		assert.Fail("Error creating test table", err)
 		return
@@ -1240,7 +1239,7 @@ func TestTriggerQueryBCustomCapture(t *testing.T) {
 		item["blob_data"] = blobData
 	}
 
-	err = populateTestTableB(db, "test_one", data)
+	err = populatePostgresSqlTestTableB(db, "test_one", data)
 	if err != nil {
 		assert.Fail("Error populating test table", err)
 		return
@@ -1287,7 +1286,7 @@ func TestTriggerQueryBCustomCapture(t *testing.T) {
 	}
 
 	trigger.Config = &modconfig.TriggerQuery{
-		Database:   "sqlite:./test_trigger_query_b.db",
+		Database:   connectionString,
 		Sql:        "select * from test_one",
 		PrimaryKey: "id",
 		Captures: map[string]*modconfig.TriggerQueryCapture{
@@ -1296,6 +1295,7 @@ func TestTriggerQueryBCustomCapture(t *testing.T) {
 		},
 	}
 
+	// Add trigger to db
 	cache.GetCache().SetWithTTL(trigger.Name(), trigger, 10*time.Minute)
 
 	executionId := util.NewExecutionId()
@@ -1312,24 +1312,16 @@ func TestTriggerQueryBCustomCapture(t *testing.T) {
 	assert.Equal(0, len(pipelineQueues), "insert capture not defined no pipeline should be executed")
 }
 
-func TestTriggerQueryWithNull(t *testing.T) {
+func TestPostgresSqlTriggerQueryWithNull(t *testing.T) {
 	ctx := context.Background()
 
 	cache.InMemoryInitialize(nil)
 
 	assert := assert.New(t)
 
-	sourceDbFilename := "./test_trigger_query.db"
-	_, err := os.Stat(sourceDbFilename)
-	if !os.IsNotExist(err) {
-		err = os.Remove(sourceDbFilename)
-		if err != nil {
-			assert.Fail("Error removing test db", err)
-			return
-		}
-	}
+	connectionString := "postgres://flowpipe:password@localhost:5432/flowpipe-test?sslmode=disable"
 
-	db, err := sql.Open("sqlite3", sourceDbFilename)
+	db, err := sql.Open("postgres", connectionString)
 	if err != nil {
 		assert.Fail("Error initializing db", err)
 		return
@@ -1354,7 +1346,13 @@ func TestTriggerQueryWithNull(t *testing.T) {
 		return
 	}
 
-	err = createTestTableA(db, "test_one")
+	err = dropPostgresSqlTestTable(db, "test_one")
+	if err != nil {
+		assert.Fail("Error deleting test table", err)
+		return
+	}
+
+	err = createPostgresSqlTestTableA(db, "test_one")
 	if err != nil {
 		assert.Fail("Error creating test table", err)
 		return
@@ -1384,7 +1382,7 @@ func TestTriggerQueryWithNull(t *testing.T) {
 		},
 	}
 
-	err = populateTestTableA(db, "test_one", data)
+	err = populatePostgresSqlTestTableA(db, "test_one", data)
 	if err != nil {
 		assert.Fail("Error populating test table", err)
 		return
@@ -1441,7 +1439,7 @@ func TestTriggerQueryWithNull(t *testing.T) {
 	}
 
 	trigger.Config = &modconfig.TriggerQuery{
-		Database:   "sqlite:./test_trigger_query.db",
+		Database:   connectionString,
 		Sql:        "select * from test_one",
 		PrimaryKey: "id",
 		Captures: map[string]*modconfig.TriggerQueryCapture{
@@ -1459,14 +1457,11 @@ func TestTriggerQueryWithNull(t *testing.T) {
 
 	triggerRunner := NewTriggerRunner(trigger, executionId, triggerExecutionId)
 
-	assert.NotNil(triggerRunner, "trigger runner should not be nil")
-
 	_, err = triggerRunner.ExecuteTriggerWithArgs(ctx, nil, nil)
 	if err != nil {
 		assert.Fail("Error executing trigger", err)
 		return
 	}
-
 	// The callback to the mocks should have been called by now
 	if generatedEvalContext == nil {
 		assert.Fail("generated eval context should not be nil")
@@ -1509,7 +1504,7 @@ func TestTriggerQueryWithNull(t *testing.T) {
 	}
 }
 
-func createTestTableA(db *sql.DB, tableName string) error {
+func createPostgresSqlTestTableA(db *sql.DB, tableName string) error {
 	createTableSQL := `create table if not exists ` + tableName + ` (id text primary key, name text, age integer, registration_date date, is_active boolean);`
 
 	slog.Info("Creating table", "sql", createTableSQL)
@@ -1521,7 +1516,19 @@ func createTestTableA(db *sql.DB, tableName string) error {
 	return nil
 }
 
-func populateTestTableA(db *sql.DB, tableName string, data []map[string]interface{}) error {
+func dropPostgresSqlTestTable(db *sql.DB, tableName string) error {
+	dropTableSql := fmt.Sprintf("DROP TABLE IF EXISTS %s;", tableName)
+
+	slog.Info("Deleting table", "sql", dropTableSql)
+	_, err := db.Exec(dropTableSql)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func populatePostgresSqlTestTableA(db *sql.DB, tableName string, data []map[string]interface{}) error {
 	// Start a transaction
 	tx, err := db.Begin()
 	if err != nil {
@@ -1529,7 +1536,7 @@ func populateTestTableA(db *sql.DB, tableName string, data []map[string]interfac
 	}
 
 	// Prepare statement for inserting into the temporary table
-	tempStmt, err := tx.Prepare(`INSERT INTO ` + tableName + ` (id, name, age, registration_date, is_active) VALUES (?, ?, ?, ?, ?)`) //nolint:gosec // should be safe to use
+	tempStmt, err := tx.Prepare(`INSERT INTO ` + tableName + ` (id, name, age, registration_date, is_active) VALUES ($1, $2, $3, $4, $5)`) //nolint:gosec // should be safe to use
 	if err != nil {
 		err2 := tx.Rollback()
 		if err2 != nil {
@@ -1562,13 +1569,13 @@ func populateTestTableA(db *sql.DB, tableName string, data []map[string]interfac
 	return nil
 }
 
-func updateTestTableA(db *sql.DB, tableName string, data map[string]interface{}) error {
+func updatePostgresSqlTestTableA(db *sql.DB, tableName string, data map[string]interface{}) error {
 	tx, err := db.Begin()
 	if err != nil {
 		return err
 	}
 
-	tempStmt, err := tx.Prepare(`UPDATE ` + tableName + ` SET name = ?, age = ?, registration_date = ?, is_active = ? WHERE id = ?`) //nolint:gosec // should be safe to use
+	tempStmt, err := tx.Prepare(`UPDATE ` + tableName + ` SET name = $1, age = $2, registration_date = $3, is_active = $4 WHERE id = $5`) //nolint:gosec // should be safe to use
 	if err != nil {
 		err2 := tx.Rollback()
 		if err2 != nil {
@@ -1596,19 +1603,29 @@ func updateTestTableA(db *sql.DB, tableName string, data map[string]interface{})
 	return nil
 }
 
-func deleteFromTestTable(db *sql.DB, tableName string, idsToDelete []any) error {
+func deleteFromPostgresSqlTestTable(db *sql.DB, tableName string, idsToDelete []any) error {
 	// Start a transaction
 	tx, err := db.Begin()
 	if err != nil {
 		return err
 	}
 
-	// Prepare statement for inserting into the temporary table
-	placeholders := strings.Join(strings.Split(strings.Repeat("?", len(idsToDelete)), ""), ",")
+	// Generate placeholders for each ID in the format PostgreSQL expects
+	placeholders := make([]string, len(idsToDelete))
+	for i := range idsToDelete {
+		placeholders[i] = fmt.Sprintf("$%d", i+1)
+	}
+	placeholdersStr := strings.Join(placeholders, ", ")
 
-	tempStmt, err := tx.Prepare(fmt.Sprintf("DELETE FROM %s WHERE id in (%s)", tableName, placeholders))
+	query := fmt.Sprintf("DELETE FROM %s WHERE id IN (%s)", tableName, placeholdersStr) //nolint:gosec // should be safe to use
+	tempStmt, err := tx.Prepare(query)
 	if err != nil {
-		return err
+		// Roll back the transaction on error
+		err2 := tx.Rollback()
+		if err2 != nil {
+			slog.Error("Error rolling back transaction", "error", err2)
+			return err
+		}
 	}
 	defer tempStmt.Close()
 
@@ -1626,14 +1643,14 @@ func deleteFromTestTable(db *sql.DB, tableName string, idsToDelete []any) error 
 	return nil
 }
 
-func createTestTableB(db *sql.DB, tableName string) error {
+func createPostgresSqlTestTableB(db *sql.DB, tableName string) error {
 	createTableSQL := `CREATE TABLE IF NOT EXISTS ` + tableName + ` (
 		id INTEGER PRIMARY KEY,  -- Changed to INTEGER and is the primary key
 		name TEXT,
 		age INTEGER,
 		registration_date DATE,
 		is_active BOOLEAN,
-		blob_data BLOB       -- for storing BLOB data
+		blob_data BYTEA          -- for storing BLOB data
 	);`
 
 	_, err := db.Exec(createTableSQL)
@@ -1644,7 +1661,7 @@ func createTestTableB(db *sql.DB, tableName string) error {
 	return nil
 }
 
-func populateTestTableB(db *sql.DB, tableName string, data []map[string]interface{}) error {
+func populatePostgresSqlTestTableB(db *sql.DB, tableName string, data []map[string]interface{}) error {
 	// Start a transaction
 	tx, err := db.Begin()
 	if err != nil {
@@ -1652,7 +1669,7 @@ func populateTestTableB(db *sql.DB, tableName string, data []map[string]interfac
 	}
 
 	// Prepare statement for inserting into the temporary table
-	tempStmt, err := tx.Prepare(`INSERT INTO ` + tableName + ` (id, name, age, registration_date, is_active, blob_data) VALUES (?, ?, ?, ?, ?, ?)`) //nolint:gosec // should be safe to use
+	tempStmt, err := tx.Prepare(fmt.Sprintf(`INSERT INTO %s (id, name, age, registration_date, is_active, blob_data) VALUES ($1, $2, $3, $4, $5, $6)`, tableName))
 	if err != nil {
 		err2 := tx.Rollback()
 		if err2 != nil {
@@ -1685,13 +1702,14 @@ func populateTestTableB(db *sql.DB, tableName string, data []map[string]interfac
 	return nil
 }
 
-func updateTestTableB(db *sql.DB, tableName string, data map[string]interface{}) error {
+func updatePostgresSqlTestTableB(db *sql.DB, tableName string, data map[string]interface{}) error {
 	tx, err := db.Begin()
 	if err != nil {
 		return err
 	}
 
-	tempStmt, err := tx.Prepare(`UPDATE ` + tableName + ` SET name = ?, age = ?, registration_date = ?, is_active = ?, blob_data = ? WHERE id = ?`) //nolint:gosec // should be safe to use
+	stmtText := fmt.Sprintf(`UPDATE %s SET name = $1, age = $2, registration_date = $3, is_active = $4, blob_data = $5 WHERE id = $6`, tableName) //nolint:gosec // should be safe to use
+	tempStmt, err := tx.Prepare(stmtText)
 	if err != nil {
 		err2 := tx.Rollback()
 		if err2 != nil {
